# public/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Blacksport",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/services/{locale?}": {
            "get": {
                "tags": [
                    "service"
                ],
                "summary": "Service Page",
                "description": "Getting services",
                "operationId": "App\\Http\\Controllers\\ServiceController::index",
                "parameters": [
                    {
                        "name": "Locale",
                        "in": "path",
                        "description": "Language",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "ru, en, uk"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "programs": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "name": {
                                                        "type": "string"
                                                    },
                                                    "description": {
                                                        "type": "string"
                                                    },
                                                    "created_at": {
                                                        "type": "string"
                                                    },
                                                    "updated_at": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/service/request": {
            "post": {
                "tags": [
                    "service"
                ],
                "summary": "Program Page",
                "description": "Create service request",
                "operationId": "App\\Http\\Controllers\\ServiceController::store",
                "requestBody": {
                    "description": "Pass params",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "service_id",
                                    "name",
                                    "phone",
                                    "email"
                                ],
                                "properties": {
                                    "service_id": {
                                        "type": "string",
                                        "example": "1"
                                    },
                                    "name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Arsen"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "380645564545"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user1@mail.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "success"
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}